refactor and clean up nodes
