+++ qualibration_graphs/quantum_dots/calibration_utils/T1/__init__.py
@@ -1,9 +1,9 @@
-from .parameters import Parameters
 from .analysis import (
     fit_raw_data,
     log_fitted_results,
     analyse_raw_data,
 )
+from .parameters import Parameters
 from .plotting import plot_raw_data_with_fit, plot_all
 
 __all__ = [
