+++ qualibration_graphs/quantum_dots/calibration_utils/ramsey_detuning/__init__.py
@@ -1,10 +1,13 @@
-from calibration_utils.T1_parity_diff.analysis import (
+from .parameters import Parameters
+from .analysis import (
     fit_raw_data,
     log_fitted_results,
     analyse_raw_data,
 )
-from calibration_utils.T1_parity_diff.parameters import Parameters
-from calibration_utils.T1_parity_diff.plotting import plot_raw_data_with_fit, plot_all
+from .plotting import (
+    plot_raw_data_with_fit,
+    plot_all,
+)
 
 __all__ = [
     "Parameters",
