+++ qualibration_graphs/quantum_dots/calibration_utils/ramsey_detuning/parameters.py
@@ -3,7 +3,6 @@ from qualibrate.core.parameters import RunnableParameters
 from qualibration_libs.parameters import (
     CommonNodeParameters,
     QubitsExperimentNodeParameters,
-    IdleTimeNodeParameters,
 )
 
 from calibration_utils.heralded_initialization_utils.parameters import HeraldedInitializeParameters
@@ -16,17 +15,21 @@ class NodeSpecificParameters(RunnableParameters):
     num_shots: int = 300
     """Number of averages to perform. Default is 100."""
 
-
 class Parameters(
     NodeParameters,
     CommonNodeParameters,
-    IdleTimeNodeParameters,
     NodeSpecificParameters,
     HeraldedInitializeParameters,
     QubitsExperimentNodeParameters,
     ParityDiffAnalysisParameters,
 ):
-    """Parameter set for 11a_ramsey."""
+    """Parameter set for 11b_ramsey_detuning."""
 
-    frequency_detuning_in_mhz: float = 0.25
-    """Frequency detuning in MHz. Default is 1.0 MHz."""
+    detuning_span_in_mhz: float = 5.0
+    """Frequency detuning span. Default 5MHz."""
+    detuning_step_in_mhz: float = 0.1
+    """Frequency detuning step. Default 0.1MHz"""
+    idle_time_ns: int = 100
+    """Short idle time in ns (gives wide fringes for coarse localisation)."""
+    idle_time_long_ns: int = 400
+    """Long idle time in ns (gives narrow fringes for precision + T2* via amplitude ratio)."""
