--- qualibration_graphs/quantum_dots/calibration_utils/ramsey_parity_diff/plotting.py
@@ -1,155 +1,0 @@
-"""Plotting for the ±δ Ramsey parity-difference analysis.
-
-Produces a two-panel figure per qubit:
-
-1. **+δ trace** — parity-difference vs idle time at positive detuning,
-   with the damped-cosine fit overlaid and fitted frequency annotated.
-2. **−δ trace** — same for the negative detuning.
-"""
-
-from __future__ import annotations
-
-from typing import Any, List
-
-import matplotlib.pyplot as plt
-import numpy as np
-import xarray as xr
-
-from calibration_utils.measurement_utils.measurement_streams import get_parity_item_names
-
-
-def _get_qubit_names_from_ds(
-    ds: xr.Dataset,
-    qubits: List[Any],
-    analysis_signal: str,
-) -> List[str]:
-    return get_parity_item_names(
-        ds,
-        analysis_signal,
-        item_names=[getattr(q, "name", f"Q{i}") for i, q in enumerate(qubits)],
-    )
-
-
-def _plot_trace_ax(
-    ax: "plt.Axes",
-    tau_ns: np.ndarray,
-    qubit_name: str,
-    trace_fit: dict | None,
-    label: str,
-    analysis_signal: str = "E_p1_given_p0_0",
-    color: str = "b",
-    fit_color: str = "r",
-) -> None:
-    """Plot one detuning trace with its damped-cosine fit."""
-    if trace_fit is None:
-        ax.text(0.5, 0.5, "No fit data", transform=ax.transAxes, ha="center")
-        ax.set_title(f"{qubit_name} — {label}")
-        return
-
-    trace = trace_fit.get("signal", trace_fit.get("pdiff"))
-    fitted = trace_fit.get("fitted_curve")
-    t_shifted = trace_fit["tau_shifted"]
-    t_plot = t_shifted + tau_ns[0]
-
-    ax.plot(t_plot, trace, f"{color}-", lw=0.8, alpha=0.7)
-    ax.scatter(t_plot, trace, c=color, s=8, alpha=0.5, zorder=3, label="Data")
-
-    if fitted is not None:
-        ax.plot(
-            t_plot, fitted, f"{fit_color}-", lw=1.5, alpha=0.9, label="Damped cosine"
-        )
-
-    ax.set_xlabel("Idle time (ns)")
-    ax.set_ylabel(analysis_signal)
-    ax.set_ylim(-0.05, 1.05)
-
-    f_hz = trace_fit.get("ramsey_freq", np.nan)
-    title = f"{qubit_name} — {label}"
-    if np.isfinite(f_hz):
-        title += f" (f={f_hz * 1e-6:.3f} MHz)"
-    ax.set_title(title)
-    ax.legend(loc="upper right", fontsize=7)
-
-
-def plot_raw_data_with_fit(
-    ds: xr.Dataset,
-    ds_fit: xr.Dataset | None,
-    qubits: List[Any],
-    fit_results: dict,
-    analysis_signal: str = "E_p1_given_p0_0",
-) -> "plt.Figure":
-    """Plot ±δ Ramsey analysis for each qubit.
-
-    Layout (per qubit row):
-    * Column 1 — +δ trace with damped-cosine fit.
-    * Column 2 — −δ trace with damped-cosine fit.
-    """
-    qubit_names = _get_qubit_names_from_ds(ds, qubits, analysis_signal)
-    if not qubit_names:
-        fig, _ = plt.subplots(figsize=(6, 4))
-        return fig
-
-    n = len(qubit_names)
-    ncol = 2
-    fig, axes = plt.subplots(n, ncol, figsize=(6 * ncol, 4 * n), squeeze=False)
-
-    tau_ns = np.asarray(ds.tau.values, dtype=float)
-    detuning_hz = np.asarray(ds.detuning.values, dtype=float)
-
-    for i, qname in enumerate(qubit_names):
-        fr = fit_results.get(qname, {})
-        diag = fr.get("_diag", {})
-        fit_plus = diag.get("fit_plus")
-        fit_minus = diag.get("fit_minus")
-
-        det_plus_mhz = detuning_hz[0] * 1e-6 if len(detuning_hz) > 0 else 0
-        det_minus_mhz = detuning_hz[1] * 1e-6 if len(detuning_hz) > 1 else 0
-
-        _plot_trace_ax(
-            axes[i, 0],
-            tau_ns,
-            qname,
-            fit_plus,
-            label=f"+δ ({det_plus_mhz:+.1f} MHz)",
-            analysis_signal=analysis_signal,
-            color="b",
-            fit_color="r",
-        )
-        _plot_trace_ax(
-            axes[i, 1],
-            tau_ns,
-            qname,
-            fit_minus,
-            label=f"−δ ({det_minus_mhz:+.1f} MHz)",
-            analysis_signal=analysis_signal,
-            color="C2",
-            fit_color="C3",
-        )
-
-    fig.suptitle(f"Ramsey ±δ triangulation ({analysis_signal})")
-    fig.tight_layout()
-    return fig
-
-
-def plot_all(
-    ds_raw: xr.Dataset,
-    qubits: List[Any],
-    *,
-    ds_fit: xr.Dataset | None = None,
-    fit_results: dict | None = None,
-    analysis_signal: str = "E_p1_given_p0_0",
-    show: bool = True,
-) -> dict[str, "plt.Figure"]:
-    """Build and return all 11a Ramsey figures."""
-    figures = {
-        "raw_data_with_fit": plot_raw_data_with_fit(
-            ds_raw,
-            ds_fit,
-            qubits,
-            fit_results or {},
-            analysis_signal=analysis_signal,
-        )
-    }
-    if show:
-        plt.show()
-    return figures
