+++ qualibration_graphs/quantum_dots/calibrations/loss_divincenzo/11a_ramsey.py
@@ -12,8 +12,8 @@ from qualang_tools.units import unit
 
 from qualibrate.core import QualibrationNode
 from quam_config import Quam
-from calibration_utils.ramsey import RamseyParameters
-from calibration_utils.ramsey_parity_diff import (
+from calibration_utils.ramsey import (
+    Parameters,
     analyse_raw_data,
     log_fitted_results,
     plot_all,
@@ -56,17 +56,17 @@ State update:
 """
 
 
-node = QualibrationNode[RamseyParameters, Quam](
+node = QualibrationNode[Parameters, Quam](
     name="11a_ramsey",
     description=description,
-    parameters=RamseyParameters(),
+    parameters=Parameters(),
 )
 
 
 # Any parameters that should change for debugging purposes only should go in here
 # These parameters are ignored when run through the GUI or as part of a graph
 @node.run_action(skip_if=node.modes.external)
-def custom_param(node: QualibrationNode[RamseyParameters, Quam]):
+def custom_param(node: QualibrationNode[Parameters, Quam]):
     """Allow the user to locally set the node parameters for debugging purposes, or execution in the Python IDE."""
     pass
 
@@ -77,20 +77,24 @@ node.machine = Quam.load()
 
 # %% {Create_QUA_program}
 @node.run_action(skip_if=node.parameters.load_data_id is not None)
-def create_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
-    """Create the sweep axes and generate the QUA program.
+def create_qua_program(node: QualibrationNode[Parameters, Quam]):
+    """Create the sweep axes and generate the QUA program from the pulse sequence and the node parameters."""
 
-    Sweeps idle time at two symmetric detunings [+δ, −δ] from the qubit
-    intermediate frequency, producing a 2-D dataset (detuning × tau).
-    """
     # ── Experiment parameters (Python side) ──────────────────────────────
+
+    # A class for unit conversion
     u = unit(coerce_to_integer=True)
 
+    # Select which qubits participate in this calibration
     node.namespace["qubits"] = qubits = get_qubits(node)
 
+    # Number of shots per detuning point
     n_avg = node.parameters.num_shots
+
+    # Extract the detuning value, and construct a small array of ±detuning
     detuning = node.parameters.frequency_detuning_in_mhz * u.MHz
     detuning_values = np.array([detuning, -detuning])
+
     # Idle time sweep (in clock cycles of 4ns)
     tau_values = get_idle_times_in_clock_cycles(node.parameters)
 
@@ -108,39 +112,51 @@ def create_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
 
     # ── QUA program (runs on the OPX in real time) ───────────────────────
     with program() as node.namespace["qua_program"]:
+        # Real-time variables:
+        # t      : QUA variable representing the idle time
+        # df     : QUA variable representing the frequency detuning
+        # n      : shot counter
+        # p_post, p_pre : post- / pre-manipulation measurement outcomes
         t = declare(int)
         df = declare(int)
         n = declare(int)
-
-        # Streams store the joint pre/post parity outcomes for later conversion into
-        # conditional expectations such as E_p1_given_p0_0.
         p_post, p_pre, streams = declare_streams(node, qubits)
-
         n_st = declare_output_stream()
 
+        # Python loop over the relevant qubits
         for qubit in qubits:
+            # Extract the current IF of the qubit's XY component
             intermediate_frequency = qubit.xy.intermediate_frequency
-            with for_(n, 0, n < n_avg, n + 1):
-                save(n, n_st)
 
+            # ── OUTER LOOP: average over shots ───────────────────────
+            with for_(n, 0, n < n_avg, n + 1):
+                save(n, n_st) # tell the PC which shot we are on
+
+                # ── MIDDLE LOOP: repeat sweep for +f and -f ───────────────────────
                 with for_(*from_array(df, detuning_values)):
+
+                    # ── INNER LOOP: sweep idle times t ───────────────────────
                     with for_each_(t, tau_values):
-                        # Reset the drive frequency and frame before each Ramsey shot.
+
+                        # Reset the drive frequency and frame before each Ramsey shot, for initialization
                         qubit.xy.update_frequency(intermediate_frequency)
                         reset_frame(qubit.xy.name)
 
+                        # Start with global align
                         align()
 
-                        # Optional pre-measurement parity readout for conditional analysis.
+                        # Optional pre-measurement parity readout for conditional analysis
                         if node.parameters.parity_measurement:
                             qubit.empty()
                             a1 = qubit.measure()
 
-                        # Re-initialize at the operating point before applying the Ramsey pulses.
+                        # Perform the initialize macro
                         qubit.initialize()
 
-                        align()
+                        # Update the frequency after the initialize macro
                         qubit.xy.update_frequency(intermediate_frequency + df)
+
+                        # Global align before and after the operate -> wait -> operate -> measure shot
                         align()
 
                         with strict_timing_():
@@ -154,25 +170,33 @@ def create_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
                         # Measure the post-sequence state and return the slow voltages to zero.
                         a2 = qubit.measure()
 
+                        # Just in-case there is any residual output, ramp everything down to zero
                         qubit.voltage_sequence.ramp_to_zero()
                         align()
 
+                        # Cast the bool output of the measurement to an int (0 or 1) for averaging purposes
                         assign(p_post, Cast.to_int(a2))
 
+                        # Optionally cast the pre-parity measurement to an int too
                         if node.parameters.parity_measurement:
                             assign(p_pre, Cast.to_int(a1))
 
+                        # Save the measurements to the relevant streams, dependent on whether node.parameters.parity_measurement
                         save_measurement(node, qubit.name, p_pre, p_post, streams)
 
+            # Reset the XY freqency to the initially stored IF
             qubit.xy.update_frequency(intermediate_frequency)
 
+        # ── Post-processing on the OPX before data reaches the PC ─────────
         with stream_processing():
             n_st.save("n")
-
             n_detuning = len(detuning_values)
             n_tau = len(tau_values)
             for qubit in qubits:
-                # Buffer one 2-D trace per qubit (detuning × tau) before host-side processing.
+                # Save order per stream: for each qubit, sweep all tau values.
+                # n_tau axis - group points along the idle time axis
+                # n_detuning - group points along the two detuning values
+                # Result: 2D joint-outcome counts vs (n_detuning, n_tau) per qubit
                 buffer_streams(node, qubit.name, streams, n_detuning, n_tau)
 
 
@@ -180,7 +204,7 @@ def create_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
 @node.run_action(
     skip_if=node.parameters.load_data_id is not None or not node.parameters.simulate
 )
-def simulate_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
+def simulate_qua_program(node: QualibrationNode[Parameters, Quam]):
     """Connect to the QOP and simulate the QUA program."""
     # Connect to the QOP
     qmm = node.machine.connect()
@@ -192,7 +216,7 @@ def simulate_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
     node.results["simulation"] = {
         "figure": fig,
         "wf_report": wf_report,
-        "samples": samples,
+        # "samples": samples,
     }
 
 
@@ -200,7 +224,7 @@ def simulate_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
 @node.run_action(
     skip_if=node.parameters.load_data_id is not None or node.parameters.simulate
 )
-def execute_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
+def execute_qua_program(node: QualibrationNode[Parameters, Quam]):
     """Connect to the QOP, execute the QUA program and fetch the raw data."""
     # Connect to the QOP
     qmm = node.machine.connect()
@@ -217,7 +241,6 @@ def execute_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
                 data_fetcher.get("n", 0),
                 node.parameters.num_shots,
                 start_time=data_fetcher.t_start,
-                node=node,
             )
         node.log(job.execution_report())
     node.results["ds_raw"] = dataset
@@ -225,7 +248,7 @@ def execute_qua_program(node: QualibrationNode[RamseyParameters, Quam]):
 
 # %% {Load_historical_data}
 @node.run_action(skip_if=node.parameters.load_data_id is None)
-def load_data(node: QualibrationNode[RamseyParameters, Quam]):
+def load_data(node: QualibrationNode[Parameters, Quam]):
     """Load a previously acquired dataset."""
     load_data_id = node.parameters.load_data_id
     node.load_from_id(node.parameters.load_data_id)
@@ -235,7 +258,7 @@ def load_data(node: QualibrationNode[RamseyParameters, Quam]):
 
 # %% {Process_raw_data}
 @node.run_action(skip_if=node.parameters.simulate)
-def process_raw_data(node: QualibrationNode[RamseyParameters, Quam]):
+def process_raw_data(node: QualibrationNode[Parameters, Quam]):
     """Compute conditional expectations from joint-outcome streams."""
     node.results["ds_raw"] = process_streams(
         node.results["ds_raw"],
@@ -246,8 +269,8 @@ def process_raw_data(node: QualibrationNode[RamseyParameters, Quam]):
 
 # %% {Analyse_data}
 @node.run_action(skip_if=node.parameters.simulate)
-def analyse_data(node: QualibrationNode[RamseyParameters, Quam]):
-    """Analyse the raw data to extract Ramsey frequency and T2*."""
+def analyse_data(node: QualibrationNode[Parameters, Quam]):
+    """Analyse the raw data."""
     node.results["ds_fit"], fit_results, node.namespace["_fit_results_full"] = analyse_raw_data(
         node.results["ds_raw"], node
     )
@@ -261,7 +284,7 @@ def analyse_data(node: QualibrationNode[RamseyParameters, Quam]):
 
 # %% {Plot_data}
 @node.run_action(skip_if=node.parameters.simulate)
-def plot_data(node: QualibrationNode[RamseyParameters, Quam]):
+def plot_data(node: QualibrationNode[Parameters, Quam]):
     """Plot the raw and fitted data."""
     fit_with_diag = node.namespace.get("_fit_results_full", node.results.get("fit_results", {}))
     node.results["figures"] = plot_all(
@@ -277,7 +300,7 @@ def plot_data(node: QualibrationNode[RamseyParameters, Quam]):
 
 # %% {Update_state}
 @node.run_action(skip_if=node.parameters.simulate)
-def update_state(node: QualibrationNode[RamseyParameters, Quam]):
+def update_state(node: QualibrationNode[Parameters, Quam]):
     """Update the relevant parameters if the qubit data analysis was successful."""
     with node.record_state_updates():
         for qubit in node.namespace["qubits"]:
@@ -290,6 +313,6 @@ def update_state(node: QualibrationNode[RamseyParameters, Quam]):
 
 # %% {Save_results}
 @node.run_action()
-def save_results(node: QualibrationNode[RamseyParameters, Quam]):
+def save_results(node: QualibrationNode[Parameters, Quam]):
     """Persist the node results and any recorded state updates."""
     node.save()
