+++ qualibration_graphs/quantum_dots/calibrations/loss_divincenzo/11b_ramsey_detuning.py
@@ -15,8 +15,8 @@ from qualang_tools.units import unit
 
 from qualibrate.core import QualibrationNode
 from quam_config import Quam
-from calibration_utils.ramsey_detuning_parity_diff import (
-    Parameters as RamseyDetuningParameters,
+from calibration_utils.ramsey_detuning import (
+    Parameters,
     analyse_raw_data,
     log_fitted_results,
     plot_all,
@@ -56,17 +56,17 @@ State update:
 """
 
 
-node = QualibrationNode[RamseyDetuningParameters, Quam](
+node = QualibrationNode[Parameters, Quam](
     name="11b_ramsey_detuning",
     description=description,
-    parameters=RamseyDetuningParameters(),
+    parameters=Parameters(),
 )
 
 
 # Any parameters that should change for debugging purposes only should go in here
 # These parameters are ignored when run through the GUI or as part of a graph
 @node.run_action(skip_if=node.modes.external)
-def custom_param(node: QualibrationNode[RamseyDetuningParameters, Quam]):
+def custom_param(node: QualibrationNode[Parameters, Quam]):
     """Allow the user to locally set the node parameters for debugging purposes, or execution in the Python IDE."""
     pass
 
@@ -77,20 +77,20 @@ node.machine = Quam.load()
 
 # %% {Create_QUA_program}
 @node.run_action(skip_if=node.parameters.load_data_id is not None)
-def create_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
-    """Create the sweep axes and generate the QUA program.
+def create_qua_program(node: QualibrationNode[Parameters, Quam]):
+    """Create the sweep axes and generate the QUA program from the pulse sequence and the node parameters."""
 
-    Sweeps drive-frequency detuning at two fixed idle times (τ_short
-    and τ_long), performing a π/2 – idle – π/2 Ramsey sequence with
-    parity-difference readout at each.  Produces a 2-D dataset
-    (tau × detuning).
-    """
     # ── Experiment parameters (Python side) ──────────────────────────────
+
+    # A class for unit conversion
     u = unit(coerce_to_integer=True)
 
+    # Select which qubits participate in this calibration
     node.namespace["qubits"] = qubits = get_qubits(node)
 
+    # Number of shots per detuning point
     n_avg = node.parameters.num_shots
+
     # Two idle times in clock cycles (4 ns each)
     idle_times_cc = np.array(
         [
@@ -98,13 +98,10 @@ def create_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
             node.parameters.idle_time_long_ns // 4,
         ]
     )
-    idle_times_ns = np.array(
-        [
-            node.parameters.idle_time_ns,
-            node.parameters.idle_time_long_ns,
-        ],
-        dtype=float,
-    )
+    # Store the ns array as a float dtype
+    idle_times_ns = idle_times_cc.astype(float) * 4
+
+    # Construct the array of frequency detunings
     detuning_values = np.arange(
         -node.parameters.detuning_span_in_mhz / 2 * u.MHz,
         node.parameters.detuning_span_in_mhz / 2 * u.MHz,
@@ -124,38 +121,51 @@ def create_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
     # ── QUA program (runs on the OPX in real time) ───────────────────────
     with program() as node.namespace["qua_program"]:
+        # Real-time variables:
+        # t      : QUA variable representing the idle time
+        # df     : QUA variable representing the frequency detuning
+        # n      : shot counter
+        # p_post, p_pre : post- / pre-manipulation measurement outcomes
         t = declare(int)
         df = declare(int)
         n = declare(int)
-
-        # Streams store the joint pre/post parity outcomes for later conversion into
-        # conditional expectations such as E_p1_given_p0_0.
         p_post, p_pre, streams = declare_streams(node, qubits)
-
         n_st = declare_output_stream()
 
+        # Python loop over the relevant qubits
         for qubit in qubits:
+            # Extract the current IF of the qubit's XY component
             intermediate_frequency = qubit.xy.intermediate_frequency
 
+            # ── OUTER LOOP: average over shots ───────────────────────
             with for_(n, 0, n < n_avg, n + 1):
-                save(n, n_st)
+                save(n, n_st) # tell the PC which shot we are on
 
+                # ── MIDDLE LOOP: sweep all detuning values ───────────────────────
                 with for_(*from_array(df, detuning_values)):
+
+                    # ── INNER LOOP: repeat sweep for idle_time_ns and idle_time_ns_long ───────────────────────
                     with for_(*from_array(t, idle_times_cc)):
-                        # Reset the drive frequency and frame before each Ramsey shot.
-                        qubit.xy.update_frequency(intermediate_frequency + df)
+
+                        # Reset the drive frequency and frame before each Ramsey shot, for initialization
+                        qubit.xy.update_frequency(intermediate_frequency)
                         reset_frame(qubit.xy.name)
 
+                        # Start with global align
                         align()
 
-                        # Optional pre-measurement parity readout for conditional analysis.
+                        # Optional pre-measurement parity readout for conditional analysis
                         if node.parameters.parity_measurement:
                             qubit.empty()
                             a1 = qubit.measure()
 
-                        # Re-initialize at the operating point before applying the Ramsey pulses.
+                        # Perform the initialize macro
                         qubit.initialize()
 
+                        # Update the frequency after the initialize macro
+                        qubit.xy.update_frequency(intermediate_frequency)
+
+                        # Global align before and after the operate -> wait -> operate -> measure shot
                         align()
 
                         with strict_timing_():
@@ -166,37 +176,44 @@ def create_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
                         align()
 
+                        # Measure the post-sequence state and return the slow voltages to zero.
                         a2 = qubit.measure()
 
-                        # Return the slow voltages to zero before the next shot.
+                        # Just in-case there is any residual output, ramp everything down to zero
                         qubit.voltage_sequence.ramp_to_zero()
-
                         align()
 
+                        # Cast the bool output of the measurement to an int (0 or 1) for averaging purposes
                         assign(p_post, Cast.to_int(a2))
 
+                        # Optionally cast the pre-parity measurement to an int too
                         if node.parameters.parity_measurement:
                             assign(p_pre, Cast.to_int(a1))
 
+                        # Save the measurements to the relevant streams, dependent on whether node.parameters.parity_measurement
                         save_measurement(node, qubit.name, p_pre, p_post, streams)
 
+            # Reset the XY freqency to the initially stored IF
             qubit.xy.update_frequency(intermediate_frequency)
 
+        # ── Post-processing on the OPX before data reaches the PC ─────────
         with stream_processing():
             n_st.save("n")
-
             n_tau = len(idle_times_cc)
             n_detuning = len(detuning_values)
             for qubit in qubits:
-                # Buffer one 2-D trace per qubit (tau × detuning) before host-side processing.
-                buffer_streams(node, qubit.name, streams, n_tau, n_detuning)
+                # Save order per stream: for each qubit, sweep all tau values.
+                # n_tau axis - group points along the two idle times axis
+                # n_detuning - group points along the detuning values
+                # Result: 2D joint-outcome counts vs (n_detuning, n_tau) per qubit
+                buffer_streams(node, qubit.name, streams, n_detuning, n_tau)
 
 
 # %% {Simulate}
 @node.run_action(
     skip_if=node.parameters.load_data_id is not None or not node.parameters.simulate
 )
-def simulate_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
+def simulate_qua_program(node: QualibrationNode[Parameters, Quam]):
     """Connect to the QOP and simulate the QUA program."""
     # Connect to the QOP
     qmm = node.machine.connect()
@@ -210,7 +227,7 @@ def simulate_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam])
     node.results["simulation"] = {
         "figure": fig,
         "wf_report": wf_report,
-        "samples": samples,
+        # "samples": samples,
     }
 
 
@@ -218,7 +235,7 @@ def simulate_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam])
 @node.run_action(
     skip_if=node.parameters.load_data_id is not None or node.parameters.simulate
 )
-def execute_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
+def execute_qua_program(node: QualibrationNode[Parameters, Quam]):
     """Connect to the QOP, execute the QUA program, and fetch raw joint-outcome data into ``ds_raw``."""
     # Connect to the QOP
     qmm = node.machine.connect()
@@ -228,14 +245,13 @@ def execute_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
     with qm_session(qmm, config, timeout=node.parameters.timeout) as qm:
         # The job is stored in the node namespace to be reused in the fetching_data run_action
         node.namespace["job"] = job = qm.execute(node.namespace["qua_program"])
-        # Display the progress bar
         data_fetcher = XarrayDataFetcher(job, node.namespace["sweep_axes"])
         for dataset in data_fetcher:
+            # Display the progress bar
             progress_counter(
                 data_fetcher.get("n", 0),
                 node.parameters.num_shots,
                 start_time=data_fetcher.t_start,
-                node=node,
             )
         # Display the execution report to expose possible runtime errors
         node.log(job.execution_report())
@@ -245,7 +261,7 @@ def execute_qua_program(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
 # %% {Load_historical_data}
 @node.run_action(skip_if=node.parameters.load_data_id is None)
-def load_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
+def load_data(node: QualibrationNode[Parameters, Quam]):
     """Load a previously acquired dataset."""
     load_data_id = node.parameters.load_data_id
     node.load_from_id(node.parameters.load_data_id)
@@ -255,7 +271,7 @@ def load_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
 # %% {Process_raw_data}
 @node.run_action(skip_if=node.parameters.simulate)
-def process_raw_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
+def process_raw_data(node: QualibrationNode[Parameters, Quam]):
     """Compute conditional expectations from joint-outcome streams."""
     node.results["ds_raw"] = process_streams(
         node.results["ds_raw"],
@@ -267,8 +283,8 @@ def process_raw_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
 # %% {Analyse_data}
 @node.run_action(skip_if=node.parameters.simulate)
-def analyse_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
-    """Fit resonance detuning via joint two-τ differential evolution."""
+def analyse_data(node: QualibrationNode[Parameters, Quam]):
+    """Analyse the raw data."""
     node.results["ds_fit"], fit_results, node.namespace["_fit_results_full"] = analyse_raw_data(
         node.results["ds_raw"], node
     )
@@ -282,8 +298,8 @@ def analyse_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
 # %% {Plot_data}
 @node.run_action(skip_if=node.parameters.simulate)
-def plot_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
-    """Plot conditional readout signal vs detuning for both τ traces with joint fit."""
+def plot_data(node: QualibrationNode[Parameters, Quam]):
+    """Plot the raw and fitted data."""
     fit_with_diag = node.namespace.get("_fit_results_full", node.results.get("fit_results", {}))
     node.results["figures"] = plot_all(
         node.results["ds_raw"],
@@ -298,8 +314,8 @@ def plot_data(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
 # %% {Update_state}
 @node.run_action(skip_if=node.parameters.simulate)
-def update_state(node: QualibrationNode[RamseyDetuningParameters, Quam]):
-    """Correct the qubit intermediate frequency by the fitted resonance offset."""
+def update_state(node: QualibrationNode[Parameters, Quam]):
+    """Update the relevant parameters if the qubit data analysis was successful."""
 
     with node.record_state_updates():
         for qubit in node.namespace["qubits"]:
@@ -317,6 +333,6 @@ def update_state(node: QualibrationNode[RamseyDetuningParameters, Quam]):
 
 # %% {Save_results}
 @node.run_action()
-def save_results(node: QualibrationNode[RamseyDetuningParameters, Quam]):
+def save_results(node: QualibrationNode[Parameters, Quam]):
     """Persist the node results and any recorded state updates."""
     node.save()
