+++ qualibration_graphs/quantum_dots/calibrations/loss_divincenzo/11c_ramsey_chevron.py
@@ -15,8 +15,8 @@ from qualang_tools.units import unit
 
 from qualibrate.core import QualibrationNode
 from quam_config import Quam
-from calibration_utils.ramsey import RamseyChevronParameters as Parameters
-from calibration_utils.ramsey_chevron_parity_diff import (
+from calibration_utils.ramsey_chevron import (
+    Parameters,
     analyse_raw_data,
     log_fitted_results,
     plot_all,
@@ -82,20 +82,29 @@ node.machine = Quam.load()
 @node.run_action(skip_if=node.parameters.load_data_id is not None)
 def create_qua_program(node: QualibrationNode[Parameters, Quam]):
     """Create the sweep axes and generate the QUA program from the pulse sequence and the node parameters."""
+
+    # ── Experiment parameters (Python side) ──────────────────────────────
+
+    # A class for unit conversion
     u = unit(coerce_to_integer=True)
 
+    # Select which qubits participate in this calibration
     node.namespace["qubits"] = qubits = get_qubits(node)
 
+    # Number of shots per detuning point
     n_avg = node.parameters.num_shots
+
     # Idle time sweep (in clock cycles of 4ns)
     tau_values = get_idle_times_in_clock_cycles(node.parameters)
-    # Detuning sweep
+
+    # Construct the array of frequency detunings
     detuning_values = np.arange(
         -node.parameters.detuning_span_in_mhz / 2 * u.MHz,
         node.parameters.detuning_span_in_mhz / 2 * u.MHz,
         node.parameters.detuning_step_in_mhz * u.MHz,
     )
 
+    # Register the sweep axes to be added to the dataset when fetching data.
     node.namespace["sweep_axes"] = {
         "qubit": xr.DataArray(qubits.get_names()),
         "detuning": xr.DataArray(
@@ -107,66 +116,93 @@ def create_qua_program(node: QualibrationNode[Parameters, Quam]):
             attrs={"long_name": "idle time", "units": "ns"},
         ),
     }
-
+    # ── QUA program (runs on the OPX in real time) ───────────────────────
     with program() as node.namespace["qua_program"]:
+        # Real-time variables:
+        # t      : QUA variable representing the idle time
+        # df     : QUA variable representing the frequency detuning
+        # n      : shot counter
+        # p_post, p_pre : post- / pre-manipulation measurement outcomes
         t = declare(int)
         df = declare(int)
         n = declare(int)
-
         p_post, p_pre, streams = declare_streams(node, qubits)
-
         n_st = declare_output_stream()
 
+        # Python loop over the relevant qubits
         for qubit in qubits:
+            # Extract the current IF of the qubit's XY component
             intermediate_frequency = qubit.xy.intermediate_frequency
 
+            # ── OUTER LOOP: average over shots ───────────────────────
             with for_(n, 0, n < n_avg, n + 1):
-                save(n, n_st)
+                save(n, n_st) # tell the PC which shot we are on
 
+                # ── MIDDLE LOOP: sweep all detuning values ───────────────────────
                 with for_(*from_array(df, detuning_values)):
+
+                    # ── INNER LOOP: sweep idle times t ───────────────────────
                     with for_each_(t, tau_values):
 
-                        qubit.xy.update_frequency(intermediate_frequency + df)
+                        # Reset the drive frequency and frame before each Ramsey shot, for initialization
+                        qubit.xy.update_frequency(intermediate_frequency)
                         reset_frame(qubit.xy.name)
 
+                        # Start with global align
                         align()
 
+                        # Optional pre-measurement parity readout for conditional analysis
                         if node.parameters.parity_measurement:
                             qubit.empty()
                             a1 = qubit.measure()
 
+                        # Perform the initialize macro
                         qubit.initialize()
 
+                        # Update the frequency after the initialize macro
+                        qubit.xy.update_frequency(intermediate_frequency + df)
+
+                        # Global align before and after the operate -> wait -> operate -> measure shot
                         align()
 
                         with strict_timing_():
+                            # Apply the Ramsey π/2 – idle – π/2 sequence for this detuning/τ point.
                             qubit.x90()
                             wait(t, qubit.xy.name)
                             qubit.x90()
 
                         align()
 
+                        # Measure the post-sequence state and return the slow voltages to zero.
                         a2 = qubit.measure()
 
+                        # Just in-case there is any residual output, ramp everything down to zero
                         qubit.voltage_sequence.ramp_to_zero()
-
                         align()
 
+                        # Cast the bool output of the measurement to an int (0 or 1) for averaging purposes
                         assign(p_post, Cast.to_int(a2))
 
+                         # Optionally cast the pre-parity measurement to an int too
                         if node.parameters.parity_measurement:
                             assign(p_pre, Cast.to_int(a1))
 
+                        # Save the measurements to the relevant streams, dependent on whether node.parameters.parity_measurement
                         save_measurement(node, qubit.name, p_pre, p_post, streams)
 
+            # Reset the XY freqency to the initially stored IF
             qubit.xy.update_frequency(intermediate_frequency)
 
+        # ── Post-processing on the OPX before data reaches the PC ─────────
         with stream_processing():
             n_st.save("n")
-
             n_detuning = len(detuning_values)
             n_tau = len(tau_values)
             for qubit in qubits:
+                # Save order per stream: for each qubit, sweep all tau values.
+                # n_tau axis - group points along the idle times axis
+                # n_detuning - group points along the detuning values axis
+                # Result: 2D joint-outcome counts vs (n_detuning, n_tau) per qubit
                 buffer_streams(node, qubit.name, streams, n_detuning, n_tau)
 
 
@@ -186,7 +222,7 @@ def simulate_qua_program(node: QualibrationNode[Parameters, Quam]):
     node.results["simulation"] = {
         "figure": fig,
         "wf_report": wf_report,
-        "samples": samples,
+        # "samples": samples,
     }
